Update README.md
This simple web application allows users to roll a virtual six-sided dice by clicking a button. Implemented using HTML, CSS, and JavaScript, it features a clean interface with a result display area and a roll button. Upon clicking the button, a random number between 1 and 6 is generated and shown in the result area. The application provides an intuitive and easy-to-use platform for simulating dice rolls.